\name{SLqPCR-package}
\alias{SLqPCR-package}
\alias{SLqPCR}
\docType{package}
\title{
  Functions for analysis of real-time quantitative PCR data at SIRS-Lab GmbH
}
\description{
Functions for analysis of real-time quantitative PCR data at SIRS-Lab GmbH
}
\details{
\tabular{ll}{
Package: \tab SLqPCR\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2007-01-02\cr
Depends: \tab R(>= 2.4.0), stats, RColorBrewer\cr
License: \tab GPL (version 2 or later)\cr
}
require(SLqPCR)
}
\author{ Dr. Matthias Kohl (SIRS-Lab GmbH) \url{http://www.sirs-lab.com}

Maintainer: Dr. Matthias Kohl  \email{kohl@sirs-lab.com}}
\references{
  Jo Vandesompele, Katleen De Preter, Filip Pattyn et al. (2002). Accurate 
  normalization of real-time quantitative RT-PCR data by geometric averaging 
  of multiple internal control genes. 
  Genome Biology 2002. 3(7):research0034.1-0034.11.
  \url{http://genomebiology.com/2002/3/7/research/0034/}
}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
